refactor : Move BuildRecreateMode enum to `jkube-kit/common`
