+++ jkube-kit/generator/api/src/main/java/org/eclipse/jkube/generator/api/GeneratorContext.java
@@ -19,6 +19,7 @@ import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.eclipse.jkube.kit.common.BuildRecreateMode;
 import org.eclipse.jkube.kit.common.JavaProject;
 import org.eclipse.jkube.kit.common.KitLogger;
 import org.eclipse.jkube.kit.config.image.build.JKubeBuildStrategy;
@@ -56,7 +57,7 @@ public class GeneratorContext {
     private String openshiftPullSecret;
     private String openshiftPushSecret;
     private String openshiftBuildOutputKind;
-    private String openshiftBuildRecreate;
+    private BuildRecreateMode openshiftBuildRecreate;
 
 
     public GeneratorMode getGeneratorMode() {
